Time: 91 ms (57.75%), Space: 42.6 MB (7.87%) - LeetHub

+++ 1508-range-sum-of-sorted-subarray-sums/1508-range-sum-of-sorted-subarray-sums.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int rangeSum(vector<int>& nums, int n, int left, int right) {
         vector<long long int> vectorSum;
-        for(long long int i = 0; i < nums.size(); i++){
-            long long int sum = 0;
-            for(long long int j = i; j < nums.size(); j++){
+        for(int i = 0; i < nums.size(); i++){
+            int sum = 0;
+            for(int j = i; j < nums.size(); j++){
                 sum += nums[j];
                 vectorSum.push_back(sum);
             }
@@ -12,9 +12,9 @@ public:
 
         sort(vectorSum.begin(), vectorSum.end());
 
-        long long int resultantSum = 0;
+        int resultantSum = 0;
         int mod = 1e9 + 7;
-        for(long long int i = left - 1; i < right; i++){
+        for(int i = left - 1; i < right; i++){
             resultantSum = (resultantSum + vectorSum[i]) % mod;
         }
